refactor: rename project from natest to datest
- Rename package directory from natest/ to datest/
- Update package name in pyproject.toml from natest to datest
- Update all CLI commands from natest to datest
- Update all documentation references (README, CONTRIBUTING, CLAUDE.md)
- Update all import statements and module references
- Update configuration files (Makefile, mkdocs.yml, .gitignore)
- Update test files and fixture comments
- Update design documents and architecture references
- Maintain all functionality while rebranding to datest

This is a comprehensive rename that preserves all existing functionality
while updating the project branding from natest to datest.
