+++ .gitignore
@@ -1,4 +1,4 @@
-# .gitignore - Natest Git Ignore Rules
+# .gitignore - Datest Git Ignore Rules
 # Copyright © 2025 Aitomatic, Inc. Licensed under the MIT License.
 
 # Python
