+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
-# Contributing to Natest
+# Contributing to Datest
 
-Thanks for your interest in contributing to Natest! This document provides guidelines for contributing to the project. Please read these guidelines before submitting a contribution.
+Thanks for your interest in contributing to Datest! This document provides guidelines for contributing to the project. Please read these guidelines before submitting a contribution.
 
 ## Code of Conduct
 
@@ -8,11 +8,11 @@ All contributors must abide by the [Code of Conduct](CODE_OF_CONDUCT.md). Please
 
 ## How to Contribute
 
-1. **Find an issue to work on:** Look at the list of open issues in the Natest repository. Pick one that interests you and that no one else is working on.
+1. **Find an issue to work on:** Look at the list of open issues in the Datest repository. Pick one that interests you and that no one else is working on.
 
 2. **Fork the repository and create a branch:** If you're not a project maintainer, you'll need to create a fork of the repository and create a branch on your fork where you can make your changes.
 
-3. **Submit a pull request:** After you've made your changes, submit a pull request to merge your branch into the main Natest repository. Be sure to link the issue you're addressing in your pull request.
+3. **Submit a pull request:** After you've made your changes, submit a pull request to merge your branch into the main Datest repository. Be sure to link the issue you're addressing in your pull request.
 
 Please ensure your contribution meets the following guidelines:
