+++ datest/__main__.py
@@ -1,11 +1,11 @@
 #!/usr/bin/env python3
 """
-Entry point for running natest as a module.
+Entry point for running datest as a module.
 
 Copyright © 2025 Aitomatic, Inc. Licensed under the MIT License.
 """
 
-from .cli import main
-
 if __name__ == "__main__":
+    from .cli import main
+
     main()
